requirements.txt
mysql_connector
pytest
